PCHR-1642: Make fabricators autoloadable
By putting the fabricators inside the source files dir, we can take advantage
of CiviCRM autoloader to make the classes available without the need for
require/include statements. This means the classes are available anywhere
(including to other extensions) as long as the extensions is installed.

Note: In the future, once we have a build process which will generate a
distributable version of CiviHR, it would be good to have a task to remove the
CRM\Hrjobcontract\Test dir, since all of its contents are supposed to be used
only during development.