PCHR-1603: Do not pro rate public holidays and remove getContractualEntitlement method
Instead making the public holidays a part of the contractual entitlement, we now
only add it to the proposed entitlement, after the pro rata is calculated. Basically, the
Pro Rata for a single contract was changed from
(CE + PH) * (WDTW/WD)
to
CE * (WDTW/WD)

And the Proposed Entitlement is:
PR + BF + PH

Where:
CE: Contractual Entitlement
PH: No. of Public Holidays
WDTW: No. of Working Days to Work
WD: No. of Working Days
PR: Pro Rata (The sum of all contracts pro rata)
BF: Brought Forward
PH: Number of public holidays between the contracts dates where "Add Public Holiday?" is set

The getContractualEntitlement was removed because there's no such information in a
multiple contract scenario. You can't just sum up the contracts entitlements to get that,
because you might end up with a huge number that doesn't reflect an actual number.
On the Entitlement Calculation page, the column which displayed this information was
also removed.

Finally, this made the __toString() method to stop working. Since there's a lot that has
to be changed on it to support the multiple contracts scenario, I just made it return an
empty string the commented out the related tests. This will be fixed in another commit

With the change, a few methods are not used anymore and were removed.