PCHR-1607: Don't allow brought forward from a period that is not over yet
When calculating the entitlement for a period the should only consider the
brought forward if there is a previous period and if it's over. The reason for
this is that, if it's not over yet, there is no way to know how much will be
left of the entitlement at the end and how much can be carried over to the
next period.