PCHR-1607: EntitlementCalculation refactoring
The EntitlementCalculation class was refactored and some of its code was extracted
to the ContractEntitlementCalculation class.

Basically, all the code that was doing calculations for an specific contract was moved
to this new class. Now, the EntitlementCalculation has an array of ContractEntitlmentCalculation
instance (one for each contract) and do the calculations for the whole period
based on these instances