PCHR-1607: Show calculation details for multiple contract
With this change, a more detailed description is displayed, including
how the pro rata is calculated for each contract and how they are
summed up at the end.

Since this is a lot more complicated than the simple string we had before,
the details are now displayed on a separate page, which is loaded via
ajax using the CRM.confirm() helper.