PCHR-1676: Don't allow saving Public Holidays with a date in the past
A public holiday cannot be saved with a date in the past anymore. This includes
changing the date of a public holiday in the future to one in the past.

A lot of tests create public holidays in the past in order to have a controlled
set of holidays, without the risk of a date falling on a working day in one year and then
falling on a weekend another year. All those started to fail right after the change. To fix
them, I added a fabricateWithoutValidation() method to the PublicHoliday fabricator,
which creates new holidays without using the create() and triggering the validation.