PCHR-1692: Move a lot of methods out of HRJobLeave
The HRJobLeave had a lot of methods not related specific to
the job leave table/entity and not event to the hrjobcontract
extension.

The recalculateAbsenceEntitlementForContact() and
recalculateAbsenceEntitlement() methods were moved to the
HRAbsenceEntitlement BAO, as they were handling data on
the absence_entitlement table.

The getAbsencePeriods() method was moved to the
HRAbsencePeriod BAO, as it was handling data on the
absence_period table.

Since the hrabsence extesion will be replace by L&A, I didn't bother
to add a lot of tests to these new methods moved to it. I added
some to the HRAbsencePeriod BAO because they were quick and
simple to do, but din't add any to the methods which recalculates
entitlements, as that would take a lot of time.