PCHR-1692: Recalculate the entitlement on the background
Recalculating the entitlement for all contacts right after saving an
absence period might take some time, so this is something that
is better to be done on the background.

Now, when saving a new period, a new job is added to the
Entitlement Recalculation queue. This queue is then consumed
by a scheduled job, which will finally execute the entitlement
recalculation.