PCHR-1195: Add a method to create LeaveBalanceChanges for a LeaveRequest
The LeaveBalanceChange::createForLeaveRequest() method was added in order
to encapsulate the logic of creating balance changes for each of the leave request
dates.

To calculate the amount of days to be deducted for a specific date, we use the
work pattern attribution to get which work pattern is active for a given user on a
given date. Next, we use that work pattern to get the amount of days to be deducted.

Currently, this doesn't support using the default work pattern on the calculation.
This happens, because the getLeaveDaysForDate() method requires a start
date, which is the work pattern attribution effective date when there is an
attribution. We don't have that information when we're dealing with the default
work pattern (because default work patterns are not "attributed" to an user),
so we'll have to think of which date should be used on that case (probably
the contract start date).