PCHR-1177: Fixed Issue!
Issue affected admin part of CiviHR also.  The problem seems to
happen because the activity's data is never queried from the DB.
Thus, the activity's type is null when the check is performed, and
causes the error when the null value is used to bring a single
result from the activitytype table.

Fixed by passing activity id instead of activity_type_id to
activityCreatedByTaskandAssignments(), so data can be loaded from
DB within function.