PCHR-1196: Fixes flaky tests
These tests would fail if executed closer to the end of the year.
By default, the Leave Period Entitlement mock return the start
and end dates as Jan 1st and Dec 31st, respectively, for the
current year. The public holidays being tested, have relative dates
like +x days. So, when executing the test at, for example,
Dec. 30th, a public holiday with a date like +4 days would fall
on the next year, outside the period dates, and the method to
fetch the balance for the period would not include the leave
request for that public holiday.

To fix this problem, the mock was updated to return start and
end dates relative to the current date