PCHR-1196: Update public holiday leave requests and creating/updating one absence type
When creating a new Absence Type with MTPHL as true, or changing that field value
for an existing one, we must update/create all the public holiday leave request for
public holidays in the future.

This must be done for all contact with active contracts during the public holiday date,
and it might take some time to process everything for companies with a large number
of employees. For that reason, this update happens in background. Whenever this
update should be executed, a new background task will be added a queue, which
will be later consumed by an scheduled job. This task, basically uses the
PublicHolidayLeaveRequest service to update everything.