PCHR-1199: Avoid creating duplicated Public Holiday Leave Requests
There is a situation where public holidays leave requests would end up
duplicated. This is how it would happen:

1. Add a new Public Holiday (a new background task would be created
to add the leave requests for everyone with an overlapping contract)

2. Create a new contract overlapping the Public Holiday created on the
previous step. This is will create a public holiday leave request for that
holiday (this has to be done before the background task created on
task 1 gets executed)

3. Execute the background task. It will create another leave request
for the same public holiday.

In order to fix this, I have updated the method which create these
requests to first check if there is an existing public holiday leave
request for the same contact and date. It only creates a new one if
there is none.