PCHR-501: Changes for default contacts
We truncate uf_match table when installing/uninstalling the extension since civicrm will automaticly
create a contact if you logged in with a user without matching contact. and in case the user email
is similar to any contact email in the civicrm side then it will be assigned by civicrm to it automaticly
in uf_match table.

We just  need to covere the case of installing the extension for the first time, So we change the default users
emails on druapl so one anyone log in with them , they will automaticly get assigned to their matching sample contacts.