PCHR-1788: Fix Advanced Search When Using Multiple Select Fields
Use of multiple select fields was sending the values selected by the user in an
array. Trim was being used on the value, assuming it was always a string. Added
check to verify if value to build query is an array.  This way, trim function
can be called on each value of the array, or on the whole value if it's a
string.