PCHR-1681: Fix errors in some of the LeaveBalanceChange BAO tests
These were flaky test which would only fail close to the end of the year.

The problem was with the relative dates of the Leave Requests and the
Absence Period. By default, the Absence Period created by the
createLeavePeriodEntitlementMockForBalanceTests method will end at
December 31th of the current year, but the leave requests have dates
like +15 days, meaning that, if the current year has less than 15 days
remaining, the leave request will be on the next period and won't be
included on the balance.