PCHR-1681: Add the LeaveRequest Service
Making the LeaveRequest BAO create its own LeaveBalanceChanges would break a
lot of tests where the assumption was that the balance changes should be created
separately. Also, in order to create the balance changes, the LeaveRequest BAO
would have to know a lot about the LeaveBalanceChange structure, which is not ideal.
For that reason, I added a LeaveRequest Service, which is responsible to create the
LeaveRequest and it's balance changes.