PCHR-1770: Add the "expired" filter to the TOILRequest.get API
This filter was implemented the same way as the special filters of the
LeaveRequest.get API. When not present or false, it returns all the
TOILRequests; when it's true, only expired TOILRequests are returned.

The TOILRequestSelect class was created, the same way as
LeaveRequestSelect class, and its is just a wrapper to Civi\API\SelectQuery.
When the "expired" flag is set, it add conditions to the where clause
to make sure only the expired requests will be returned