PCHR-1787: Fix Search by Contract Type in Advanced Search
Search by contract type can be done by selecting multiple options.  Array of values selected by user
for search was being imploded into a string of the form ('value1', 'value2', ... 'valuen') and then
passed to CRM_Contact_BAO_Query::buildClause(), which caused the value to be treated as a single
string, and the clause to be built as: IN ("\'value1\', \'value2\', ... , \'valuen\' ").

Fixed by passing value as an array to CRM_Contact_BAO_Query::buildClause(), which takes care of
building the clause and quoting string values apprropriately.