PCHR-1845: Fix more flaky tests
A few tests related to ContactWorkPatterns started to fail
at the beginning of 2017. This was due to how the effective
date was set on the fabricator. It used date('Y-01-01'),
which worked for the tests created in 2016, because they
were all created after 2016-01-01. In other words, some
tests assumed the effective date would always be in the
past. Now that we have a new year, date will return
2017-01-01 and some tests have fixed dates in 2016,
meaning that no ContactWorkPattern will be effective on
that date. To fix it, I changed the fabricator to have a
hardcoded date in the past (2010).

On a few places, instead of relying on the default date
set by the fabricator, I opted for passing one specific
date that would work for the test

I also fixed a problem on which the created absence period
would not cover the leave requests on the test and would
make LeaveBalanceChange return the wrong amount of days
as the balance.