PCHR-1123: Group balance changes by type in LeavePeriodEntitlement.getBreakdown
Balance changes of type "Overridden" are groupped together with
balance changes of type "Leave". "Overridden" balance changes are
just an internal way of keeping track of overridden values for the
Leave.

Now that this groupping happens, a few tests were updated in order to
reflect this new behavior. Tests where multiple balance changes of
the same type were created, were updated to create only on balance
change of each. There are new tests to test the groupping specifically.