PCHR-1265: Move the logic inside LeaveRequest.getBalanceChangeByAbsenceType to the API function
It didn't make sense to keep that method inside the LeaveRequest BAO,
since it's actually only calls methods on LeaveBalanceChange and
LeavePeriodEntitlement and doesn't really use anything from the
LeaveRequest BAO. Since the method was only used by the API with
the same name, it made sense to move the logic there.

With that, a couple of tests duplicated between the API and the BAO method
were removed and some of the tests that were only on the BAO level were
moved to the API tests.