PCHR-1894: Improvements to the LeaveRequestStatusMatrix tests
The tests for leave approvers and admins were using a data provider that,
besided returning the statuses for the tests, it also create the users.
This means that a pair of contact + manager was created each time the
provider was called. I moved the contact creation part to the test, so
this ways the same contacts will be used for all the tests.

This also fixed a problem with the tests failing when running all of
them on a just created db. I couldn't find the reason for this failure
though.