PCHR-1833: Fix notices in the LeaveRequest Service tests
The data provider was using a class property (leaveRequestStatuses) that is only set on
the setUp method, but the dataProviders are fetched BEFORE the setUp method gets
called, so array of statuses would be empty.  This was causing a lot of "index not defined"
notices because we were trying to access the statuses in an empty array.