PCHR-1681: Fix handling of to_date_type and from_date_type on calculatebalancechange
The fields, originally named to_type and from_type were renamed to to_date_type and from_date_type. The reasons are:
- Keep consistency with the field names from the LeaveRequest
- Make sure the API layer is capable of properly converting names (like half_day_am and half_day_pm) to the right option value values
- Make sure the API layer can validate that only valid options are passed to the API

Also, the calculateBalanceChange method on the LeaveRequest BAO was fixed to handle those fields values properly. Before, it was expecting to receive the option values names instead of their values.