PCHR-1858: Add Tests for Insurance Plan Type Option Values
Added assertions to test for use of option value labels when importing
contracts. Added option the define value on OptionValue Fabricator. Tests were
failing due to an error caused on HRJobContract BAO class, which was using the
function "module_exists" to test if rules drupal module was loaded. This
function is not available on unit testing environments.