PCHR-1858: Implement Insurance Plan Type as Option Value on Import
Insurance Plan Type was changed to option value to be managed by users, but
still relied on methods built for hard-coded options in HRJobContract module.
Implemented use of Option Value related methods.

Also, added a negative test for import API, expecting an error on import of an
insurance plan type that doesn't exist.