PCHR-459: Fixing hrcase installing errors
Because managed entity handler runs after upgrader install method (in which we where running our upgraders) , and because managed entity handler is not clever enough
to not trying to create a case type if it is not already exsits and because we already inside one of the upgraders (1402) create all case types then the managed entity
handler will also do the same causing errors when installing hrcase extension.

Here I just moved the call to upgraders to postInstall instead of install to ensure managed entity handler run before our upgrader and in 1402 upgrader I ensured that
it will not try to create duplicate case type but update the exsiting one instead.