PCHR-1893: Include Future Revisions in Search of Current Contracts
Future revisions were being treated as past contracts.  Since there isn't a
list of future contracts, the decision was to show future contracts as current.

Added logic to _civicrm_hrjobcontract_api3_get_current_revision so that if no
current revision is found, it searches for a future revision.

Included clauses to query on HRJobContract::getCurrentContract() to take into
account revisions in the future and limited to current contracts.

Added an assertion to test use of HRJobContract::getCurrentContract() does not
return null when there are current contracts defined.