PCHR-2050: Add/Remove Sample data with sql scripts
The reason for this is to help with the setup of the test site
and developer sites. With it, we can have a site ready with the
sample data right after installing the extension.

This is useful only during developement, though, and it should be
removed before the released of the first version.