PCHR-1711: Put Currency and Amount in a Single Row
Finished implementing required updates for the PayScale form, which required
putting currency and amount fields in the same row and removing label for
currnecy.  Also, fixed upgrader which was deleting the contents of pay_scale
field for the "Not Applicable" option.

Removed option to edit "Not Applicable" pay scale also.