PCHR-2082: Remove default ACLs for Comments
Using hook_civicrm_selectWhereClause, we remove all the default ACLs from the
Comments queries. We do this so only the more specific ACLs (added using
hook_hrcomments_selectWhereClause) get added to the query.

Without doing that, the users might not be able to see all the comments they should be.