PCHR-1834: Change order on which the extensions are installed in tests
Since l&a depends on all the other extensions (hrcore, hrjobcontract and hrcomments),
we need to install them first in order to avoid errors.

More specifically, this avoids an error caused by the sample data in the auto_install.sql file.
In the future, when we start using upgraders for the extension, this might avoid errors where
something in one upgrader depends on something else in one of the other extensions.