PCHR-1834: Staff member can change status to "waiting approval"
Whenever the staff member updates a non approved/cancelled/rejected leave request, the
status should be moved back to "Waiting Approval". There were two conflicting checks
about this in the system:

- LeaveRequestStatusMatrix: would return that this status change is allowed
- LeaveRequestRights: would return that this status change is not allowed

One possible fix would be changing LeaveRequestRights to follow the same rules as
LeaveRequestStatusMatrix. In fact, this was the first thing I did, only to realize
we would have the same logic duplicated between the two classes. For this reason,
I removed all the checks related to statuses from LeaveRequestRights and the only
class responsible for this now is LeaveRequestStatusMatrix.

The LeaveRequestRights is still responsible for non status related changes (for example,
to check which user can change the dates of a request).