PCHR-1406: Make 0 the Deefault Value for FTE Values
Default values for FTE in database was 1, for both nnumerator and denominator
of FTE.  This caused an FTE of 1 to be calculated on import for contracts with
an empty satndard hours value.

Fixed by altering contract hours table so default values are zero and
implemented default handling of FTE values on import.

Also added an assertion to FTE tests that verifies import of only mandatory
fields for contract fets a 0 FTE, Numerator, and Denominator.