PCHR-1687: Only Delete Values on Upgrade if Not Used Elsewhere
Requirement was to delete options that should not appear by default on CiviHR.
However, these values may be used in existing installations, so they should be
deleted only if it is made certain they are not used within the installation
being upgraded.

Added checks prior to deletion of values for hrjobcontract and hrcore
extension upgraders.