PCHR-2058: Remove old entitlement calculation when updating leave on a contract
Whenever the leave information on a contract was updated, the entitlement
for the contract period was recalculated. This was done using the
recalculateAbsenceEntitlement method of the CRM_HRAbsence_BAO_HRAbsenceEntitlement
BAO, from the hrabsence extension.

This recalculation isn't necessary anymore, because now, after updating a
contract, the user is redirected to the Entitlement Calculation page.

The piece of code has to be removed though, because it will cause errors
if we try to save/create a contract while the hrabsence extension is
disabled.