PCHR-1820: Show Selected Pension Provider When Editing Contract
Using "View Current Revision" or "Correct Error" buttons on a contract, and
navigating to Pension Tab, always showed Pension Provider empty, even though a
provider had been selected and saved, and currently shown on contract summary.

The problem was due to the use of an uncapitalized 'pension_provider' key to
reference the contact in the controller for the form. Fixed by changing
'pension_provider' to 'Pension_Provider'.