PCHR-1687: Move Job Enabling to Extensions
Moved enabling / disabling of scheduled jobs to the enable / disable methods of their
corresponding extensions, this way jobs will be enabled/disabled with the extensions
they depend on.

Also removed deletion of values on upgraders, and deleted instead where the vlues where
being created on installation.  This way, new installations will not have the values, but
existing installations will be left untouched.