PCHR-1687: Update CSS Path Variable Name and Doc-block for Core Jobs
In hrui extension, Upgrade Step 4701, $bootstrapcivicrmDirectory was not a
good variable name, now that bootstrap extension has been replaced by
shoreditch. Plus, it was not appropriately camelCased. Changed name to
$customCSSPath.

On the other hend, doc-block for scheduled jobs array to be enabled/disabled
in HRCore needed to be updated, now that two different arrays are used for
jobs to be enabled vs jobs to be disabled.