PCHR-2113: Fix warning on the Activity.getAbsences API
GetAbsences is a custom action added to the Activity API.
It uses the _civicrm_api3_activity_get_formatResult function,
which had its signature changed by this commit:
https://github.com/civicrm/civicrm-core/commit/bc4b6f0f5694d1a5b8e2151e6dcf9a394c300df4#diff-7673c7d055caeab2a3ec5d05a1553293R395

Due to this change, any call to this action was resulting in a
warning, because GetAbsences was not passing the new third param
added to the function.

This fixes the warning by fetching the options from the $params
array and passing it to the function.