PCHR-1687: Rename Upgrader Methods
Staging branch already had an upgrade_1027 and uprade_1028 method. When
rebasing to solve the conflict, staging's upgrade_1027 was renamed to
upgrade_1029.

Now, on installation using staging branch, upgrade_1027 was executed. When
branch for PCHR-1687 is checked out, new upgrade_1027 is not executed.

Fixed by renaming upgrade_1027 to upgrade_1029 and leaving old upgrade_1027.