PCHR-1687: Catch Exception When Creating PK for Hours Location Table
Staging had already been fixed, adding primary key and automuneric constraints
to id field of civicrm_hrhours_location table.  This caused an exception to be
thrown, which interrupted the upgrading procedure.

Fixed by catching the exception, and throwing it only if primary key has not
been set for the table.