PCHR-2271: Searches for drupal.make file locally instead of remotely
Depending on github to always download the file it's not very safe.
Specially, because repositories can change URL and branch structure. The
drupal.make file should always be bundled with the script.

I also removed all the code used to replace variables in the make file,
since we're not using any variables in it