PCHR-1711: Fix HRJobContract Upgrader 1302 Error
Upgrader steps 1032 and 1031 were both adding a primary key constraint to
hours location table. This caused an exception to be thrown, when the system
tried to add the constraint for a second time.

This error also stopped the rest of the upgrades to be run, which left the
system in an unstable state, causing the errors reported when trying to update
a payscale. This was due to the fact that the name for the periodicity column
in civicrm_hrpay_scale table was not changed to the expected pay_frequency.