PCHR-1711: Fix JS Error Shown on Pay Scale List
Pay Scale list should allow users to edit some of the values in-place. To do
this, CiviCRM has a widget that implements this behaviour, that used to be
initialized making a JS call to .crmEditable().

As of CiviCRM v4.6 onward, call to CRM.$('.crm-editable').crmEditable(); is no
longer needed, as it is done automatically on every page load. Some templates
of CiviHR are still making the call, generating a JS error.

Fixed by deleting the offending code on the template.