PCHR-2154: Calculate Hours Amount for Full Time Hours Type
On import, if Hours Type corresponds to Full Time, then the value of
hours_amount should be copied from the standard_hours value of the given
hours_location, even if a different value for hours_amount is given by the
user.

Implemented the assignment of standard_hours to hours_amount only if
hours_type is 'Full_Time'.