PCHR-2381: Update LeaveBalanceChange::getLeaveRequestBalanceForEntitlement to only consider balance changes from requests overlapping contracts
Due to this change, multiple tests had to be updated to also include a contract
covering the dates being tested, otherwise the expected balance changes would not be counted.

One test related to Public Holidays already had a contract, but it was relying on the old
method behavior and using getLeaveRequestBalanceForEntitlement to check if requests were created
(checking if the balance summing all the requests) to count requests not overlapping
contracts. It was updated to check for the actual requests instead of their balance.