PCHR-2495: Fix intermittent failures in LeaveBalanceChange BAO tests
Some tests were using the findToilRequestMainBalanceChange() helper method,
which would fail intermittently for tests using TOIL requests with a single
leave Request date.

The "Main Balance Change" for a TOIL request is balance change linked to
the first date of the request. It's the main one, because it's where we
store the accrued amount and the expiry date.

When we expire a TOIL Request, a second Balance Change is linked to the
date, with the amount of expired days. Since both are linked to the same
date, and the helper query order the results by date, it's not possible
to always get the balance changes in the same order. Sometimes the
expired balance change will return first, and the tests will fail, and
sometimes the main one will be returned, making the tests pass.

To avoid that, the query was fixed and not it searches only for balance
changes where the expired_balance_change_id is null, which eliminates
the changes of getting an expired balance change.