PCHR-2608: Make the sample data more dynamic
The IDs of the civihr_staff and civihr_manager contacts can vary
depending on a number of factors:

- Installing the CiviCRM sample on the system
- Having or not the civihr_local_admin user
- Any other factor that might end adding contacts to the db before we
create the civihr_staff and civihr_manager users

For this reason, their IDs cannot be hardcoded the way they were before.

This commit also includes a small optimization on the cleanup script
regarding contracts. This is now done with a single query and we only
delete the contracts related to civihr_staff.