PCHR-2427: Fix failing HRJobRoles tests
There were 2 errors. The first one was related to how getCurrentDepartmentsList
was dealing with roles without an end_date. When a Job Roles doesn't have a
end_date, its value is set to null, but the query was considering that the value
would be 0.

The second one was happening because getCurrentDepartmentsList considers only
job roles with a start date, but the fixtures were created without one.