PCHR-2789: Don't cache the Civi Container while running the L&A tests
L&A adds some services to the Civi Container, but if the test suite that
was run before didn't install L&A, those services won't be included in
the Container Cache and the L&A tests that use them will fail. By
disabling the Container Cache, we make sure we always start from a
clean state.