PCHR-2789: Fix undefined index notices on LeaveRequestRights tests
One of the data providers in this class was still using the old format
of data returned by getLeaveRequestStatuses, where the ID of the status
would be stored in an 'id' property.

Even with this problem, the tests were still passing. The reason was
that the setUp() method was overriding the internal array used by the
trait to cache the statuses with the correct values. To avoid having the
tests only passing by an accident a did a few changes:
- Don't overwrite the status in the setUp() method anymore
- Updated the tests that were accessing the `leaveRequestStatuses`
property directly to use dataProviders instead