PCHR-3374: Changes to use path instead of item name
I realized (too late) that civicrm_navigation does not use any unique
constraint on menu item name, so a lot of the functions I had prepared
would be unpredictable when the item name was duplicated. I changed all
functions to use path instead of item name. Path is still not unique,
but it's much less likely to be duplicated (having the same item name in
the same submenu doesn't really make sense)