PCHR-3429: Remove "Mailings" from the main navigation menu
Until CiviCRM 4.7.32, "Mailings" was only visible in the menu if the
current user had the "access CiviMail" permission. In CiviHR, the CiviMail
component is disabled and so no user had such permission.

On CiviCRM 5.0, they changed the permissions required to view the menu.
Now, it's visible to either those who have the "access CiviMail"
permission or the "send SMS" permission. The "send SMS" permission
does not depend on the CiviMail component being enabled, which means
that super users will always have it and the "Mailings" menu will be
visible. This PR hides it using hook_civicrm_navigationMenu()