PCHR-3429: Fix group by error on MySQL 5.7
The query used by the WorkPattern.findWithNumberOfWeeksAndHours() method
would not work in databases where the ONLY_FULL_GROUP_BY option is enabled
(which is the case in MySQL 5.7). The reason is that it uses a GROUP BY
clause that does not include all the fields listed in the SELECT clause.

Listing all the fields in the group by manually is not a good idea, as we
would need to update the query every time a new fields is added or
removed from the entity. So, to avoid the error, the method was removed
and now we have two additional methods to get the number of weeks and
hours for a Work Pattern. Clearly, this is not as performant as the
previous version, but it is still good enough because:
 - The new methods work based on some cached data in the entity
 - The methods are only used in the Work Pattern list page, which is not
 accessed often and system don't usually have many Work Patterns anyway