PCHR-3429: Avoid duplicated Option Value error when installing hrjobcontracts
The Employee_Permanent option value is created when we import the option_group_install.xml
file during the installation and code in the upgradeBundle() method was duplicating it.

In previous version of CiviCRM, this would fail silently, but in since 4.7.28 (because of
this PR https://github.com/civicrm/civicrm-core/pull/11089) it now not possible aymore to
have more than one Option Value for the same Option Group with the same value, and this
was throwing an error during the installation.