PCHR-3429: Avoid error when enabling the GBP currency
This upgrader does a few things regarding currencies. One of these things is
to make GBP the default currency. Another one is to enable the currency itself.

In previous versions of Civi, setting a currency as default would not enabled it,
and this is why this was done here in two steps. However, something changed
between Civi 4.7.27 and 5.3.0 and now the currency is automatically enabled when
set as the default one. Sadly, if you try to enable a currency that is
already enabled, it will give you an error. For this reason, this PR removes
the line that enabled the currency and keeps only the one that sets it as
default.

I could not find the exact change in Civi that resulted in this error.