PCHR-3681: Remove unnecessary store param in calls to CRM_Utils_Request::retrieve()
The main reason for this is that `$this` was being passed as the store param
in places that we didn't have a context object (i.e. outside of a instance
method). Up to PHP 7.0, this would work without a problem, but starting
from PHP 7.1, this now results in an error: https://3v4l.org/QXYLB

In a few places, instead of $this, `CRM_Core_DAO::$_nullObject` was being
used. This is a risky approach, as it's possible for $_nullObject to not
be null (it is just an public static property of the DAO class). This
has been removed as well