PCHR-4473: Fix custom icons of the contact summary page tabs
Since CiviCRM 5.8, the tabs now have icons by default. Some of them are
different from the ones used in CiviHR so we have to customize them to
the ones we want.

The tabset hook was used to set the custom icons. For icons in core tabs,
the hrcore does the customization. For the other tabs, the customization
was done in the extension the tab belongs to.

As part of the change in core to support those icons, they also changed
the markup of the tab and now the label/title is wrapped around a `span`
tag. This change broke some javascript code in the hrui extension that
was renaming the title of the summary tab to "Personal Details", as the
jQuery code used there was not aware of the span and was replacing the
entire tab text instead of only the span, resulting in it removing the
icon. This was fixed by using the tabset in hrcore to replace the text.
This way, even if the markup changes again in the future, the replacement
will still work