IN-117: Update environment variables to support newer versions of Jenkins plugins
As part of a Jenkins upgrade and plugins reconfiguration some environments
variables stopped being present or had their name changed. Due to that, the
Jenkinsfile had to be updated and old variables have been replaced to use
the new ones.

The GIT_URL variable was removed and there is no equivalent new one, so in this
case, on the environment session, we declare a new GITHUB_REPO_URL variable which
is used in place of the old one.